#9
跟着UP主，学会了创建BOSS，清空OBJECT， 进一步了解了 HANDLER 和 RENDER 在这种游戏中起到的不同作用

+++ src/learnaGame/GameObject.java
@@ -4,12 +4,12 @@ import java.awt.Graphics;
 import java.awt.Rectangle;
 
 public abstract class GameObject {
-	protected int x,y;
+	protected float x,y;
 	protected ID id;
-	protected int velX,velY;
+	protected float velX,velY;
 	
 	
-	public GameObject(int x, int y,ID id) {
+	public GameObject(float x, float y,ID id) {
 		this.x=x;
 		this.y=y;
 		this.id=id;
@@ -23,16 +23,16 @@ public abstract class GameObject {
 	 
 	
 	
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x=x;
 	}
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y=y;
 	}
-	public int getX() {
+	public float getX() {
 		return x;
 	}
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 	
@@ -44,19 +44,19 @@ public abstract class GameObject {
 		return id;
 	}
 	
-	public void setVelX(int velX) {
+	public void setVelX(float velX) {
 		this.velX=velX;
 	}
 	
-	public void setVelY(int velY) {
+	public void setVelY(float velY) {
 		this.velY=velY;
 	}
 	
-	public int getVelY() {
+	public float getVelY() {
 		return velY;
 	}
 	
-	public int getVelX() {
+	public float getVelX() {
 		return velX;
 	}
