+++ src/learnaGame/Trail.java
@@ -41,7 +41,7 @@ public class Trail extends GameObject {
 		Graphics2D graphics2d=(Graphics2D) g;
 		graphics2d.setComposite(makeTransparent(alpha));
 		g.setColor(color);
-		g.fillRect(x, y, width, height);
+		g.fillRect((int)x, (int)y, width, height);
 		graphics2d.setComposite(makeTransparent(1));
 
 	}
